CRAS: add -Wno-unused-function when HAVE_FUZZER
We noticed that some builds will fail when |HAVE_FUZZER| is defined,
because some functions are substituted with an alternative
implementation that would make some function unused.

In this CL, we modify the compile flags so that the check is skipped
exclusively under that flag.

BUG=b:259420892
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I6ff639dbecb38845056469fb58d3d5a7d7f04442
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4121759
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>