CRAS: Modify InputNodeGainChanged to send percentage gain
Currently, the InputNodeGainChanged signal sends dbfs*100 as the gain.
However, cras_audio_handler in Chromium uses gain in percentage from 0
to 100 for the node gain changes and updating UI.

When cras_audio_handler receives the dbus signal info, it can't make use
of the gain in dbfs. cras_audio_handler does not have info on the max
dbfs for the node, and can't turn the dbfs into percentage of max gain.
This results in issues regarding updating the gain and UI through dbus
signal triggers.

Modify the signal to send the gain in (percentage of max gain) * 100

BUG=b:255278344
TEST=`dbus-monitor --system "type='signal',sender='org.chromium.cras'"`

Cq-Depend: chromium:4125078
Change-Id: I8f82f2ef7d417f471d421a3b12f5b0e1c8d3271f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4125095
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Baili Deng <bailideng@google.com>
Tested-by: Baili Deng <bailideng@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>