CRAS: rate-limit SEVERE_UNDERRUN notification events
There is not much of a point not limiting SEVERE_UNDERRUN notification
events since it is meant to be triggered when the buffer has been off
for 5s worth of frames of data, and will be scheduled to be reopened
upon that.

In this CL, we rate-limit SEVERE_UNDERRUN events the similar to that
of UNDERRUN events.

BUG=b:263313817
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: I6febc80259df2c09141a3a4b99885945d2c3a01e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4123922
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>