CRAS: server_stream - Add new type
server_stream used to only support a single stream for echo reference.
Add another stream type for VAD, which is required to "host" the
WebRTC-APM when a RTC stream is not using AEC.

The SERVER_STREAM_CLIENT_ID is changed to 1 so the echo reference
server stream has a non-zero stream ID, making it easier to tell from
the state of no stream in logs.

BUG=b:262832006
TEST=AEC still works

Change-Id: I0ca61b317785d0b78e418b313d31c96a0db9f531
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/3990486
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>