apm: make code clearly infallible
The switch here is only used if the if statement reads out the data. In
theory we should never reach it without the read being done, but static
analysis tools don't assume that. Move this to satisfy and also protect
against possible future modification

BUG=None
TEST=None

Change-Id: I13501e0564f7795730ad578109ca7a4df96f2a6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4206131
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>