CRAS: Remove reconnect while POLLHUP
The reconnect logic causes the system to be in a weird state that
polling SCO always timeouts. It can only be recovered by reboot.

This is observed with Pixel Buds Pro but not Pixel Buds A. The repro
steps are:

1. Start recording
2. Connect Pixel Buds Pro
3. Turn it off
4. Turn it on (It should auto connect)

Revert the logic as a fast mitigation, considering the issue
crrev.com/c/3840988 tried to address is recoverable.

BUG=b:251739875
TEST=Verify with Pixel Buds Pro that SCO connect works

Change-Id: Ib21433ffdd3bd76fdbcc632bab3dd75d9f2d5ae7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4195391
Reviewed-by: Eddy Hsu <eddyhsu@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>