ucm-config: Force Hardware Volume for Logi Dock
The Logi Dock reports an abnormal volume range that triggers the SW
volume fallback, which makes the volume recorded by the device itself
out of sync with the ChromeOS system volume.

Forcing it to use the hardware volume to solve the out of sync issue.

BUG=b:266927900
TEST=flash and play with Logi Dock and other typical USB headsets

Change-Id: If511d37b0034ef5cd6d4e2b0896ce5681aac173d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4235658
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Li-Yu Yu <aaronyu@google.com>