cras.sh: Allow overriding ALSA loopback config in tests
We need to update loopback parameters in test, however,
when running tests in lab the the rootfs may be readonly.

Furthermore the the minijail bind mount does not propagate, so
we add an explicit bind mount of the HiFi.conf we want to override.

BUG=b:265748347
TEST=mount --bind /dev/null /usr/share/alsa/ucm/Loopback/HiFi.conf
TEST=minijail0 -P /mnt/empty -b /,/ -b '/usr/share/alsa/ucm/Loopback/HiFi.conf,/usr/share/alsa/ucm/Loopback/HiFi.conf' -- /bin/cat /usr/share/alsa/ucm/Loopback/HiFi.conf
TEST=see empty output above

Change-Id: Ie0a80d868008bfd7d14314cc04e26abd44513f3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4263609
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>