libcras: implement BufferCommit::latency_bytes and AsyncBufferCommit::latency_bytes
Implement latency_bytes() for CrasPlaybackData and CrasCaptureData to
allow virtio sound device to report the audio latency. This CL converts
the the ts field in the cras_audio_shm_header to provide the
implementaion of latency_bytes().

For playback, the cras_audio_shm_header::ts means the time that the
first sample in the PlaybackBuffer will be played. For capture,
cras_audio_shm_header::ts is the timestamp of the first sample at the
CaptureBuffer was recorded.

BUG=b:269989582
BUG=b:274965705
TEST=play Audio Video Sync Test video from the ARCVM youtube app.

Change-Id: Ic1faefb5ccd6186c2cc0435a1dd9ad5dfb79052e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4273506
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Tested-by: Judy Hsiao <judyhsiao@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>