Fix memory bug for softvol curve
The softvol_get_scaler function currently does not check
whether the volume is within the valid range of 0-100.
If the volume is greater than 100,
this can cause the CRAS service to crash.
To fix this issue, the function now returns the
maximum scaler when the volume is greater than 100.
Additionally, the function now logs this unexpected behavior to help with debugging.

BUG=chromium:1420631
TEST=Unit test

Change-Id: I3a56c79145f479207456d88c8a94d0e28106a660
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4306497
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Ching Yun Chang <whalechang@google.com>