CRAS: Set client type on the client
Added cras_client_set_client_type to allow setting client type on the
client before adding any stream. This is needed to allow iodev filtering
on the client side based on client type. There will be iodevs internal
to CRAS for multiple endpoints. They should be visible only to the test
client for logging and debugging.

This patch is a no-op for existing clients. It keeps the existing
behavior of cras_client_stream_params_set_client_type which is marked as
deprecated in favor of setting client type on the client itself. The
client type is not expected to change frequently, if ever, during the
lifetime of a client.

cras_test_client is modified to use the new API. To make testing easier
the --client_type argument can be either a number or an unique substring
of a client type enum name, e.g. test/chrome/lacros/crosvm.

BUG=b:250544745
TEST=Set and verify client types through cras_test_client.

Change-Id: I60ac98b62c01418ea89d3f14674e887e0b7cd71c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4346155
Commit-Queue: Ben Zhang <benzh@chromium.org>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>