Floss: Enable HfpScoConnectionError
The meaning for the HfpScoConnectionError is actually different for
BlueZ and Floss.

For BlueZ, CRAS connects to the fd directly talk to the controller.
For Floss, CRAS connects to the fd that talks to Floss instead.

We'll need to refine this UMA to better reflect the meaning of this UMA
but enabling it first so that we can compare these two stacks from audio
stack's perspective.

BUG=b:239978865
TEST=build and unit test.

Change-Id: I72104226f11f174e67778777eae79f1ff74856de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4349807
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Tested-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Auto-Submit: En-Shuo Hsu <enshuo@chromium.org>