CRAS: suppress warning when fallback due to profile-switch
When there is a profile-switch, there will be a moment where there is no
active device, which is of general interest to note of but not in this
case.

In this CL, we suppress warnings produced by profile-switch as they
are part of the normal process.

BUG=b:274914956
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: Ic32901d67906a23fdab71985aa933f53d5cd3b65
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4369162
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>