bazel: Less duplication on specifying Rust deps
We are a single cargo workspace now, so we can use the manifests
feature of crates_repository.

This allows us to use all_crate_deps(), `manifests =`, making the
Cargo.toml files to be the single source of truth for dependencies.
And we can get rid of the manually written crate_deps.bzl.

BUG=b:274602768
TEST=CARGO_BAZEL_REPIN=true bazel sync --only=crate_index

Change-Id: I99633d056e8c8b9b8768df941859fc621c014c20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4408285
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>