Replace cros_async with Tokio
The audio_streams API provides an abstraction for the async executor
used for async audio. We can use this to provide an alternative
implementation using tokio.

This allows us to get rid of cros_async usage in ChromeOS. The new
TokioExecutor is used by libcras to implement the sync AudioSocket
API and to run some tests.

BUG=b:279800844
TEST=cargo test && CQ

Change-Id: If4db72e4f6e99418ec150818f9b4e7cc971541a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4616195
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: Dennis Kempin <denniskempin@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>