CRAS: Add board config for speaker output latency offset
The reported audio output latency from CRAS is not always accurate due
to errors such as components like smart amp not properly reporting
latency and SOF not accounting for additional buffers. This can cause
issues for clients that rely on accurate audio latency values, such as
audio mesh.

Some fixes to properly report the mentioned unaccounted latency are in
the long-term, and not all sources of unaccounted latency have been
identified. However, an offset can be added based on the error between
the reported and measured latency to ensure that the reported output
latency is more accurate.

Add a board config for speaker output latency offset to ensure that the
reported output latency is accurate.
The speaker is the only peripheral that has currently available data
on measured output latency, so it is the only one that has offset
currently being added.

BUG=b:286998321
TEST=emerge adhd and deploy to device
TEST=Add 'speaker_output_latency_offset_ms' to board.ini and check that
cras_client reported latency for speaker is affected, and that headphone
latency is not affected.

Change-Id: I88423f8abcfb784446577fa3a8a35299796ecc53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4622969
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Baili Deng <bailideng@google.com>