CRAS: rework |set_audio_config| to config codec type
We used to only support the mandatory codec, thus we never configured
the codec type.

In this CL, we allow to use the reworked API that could specify the
codec type. Note that the flag is disabled by default and thus the
specified codec type should still be the mandatory codec.

BUG=b:268424393
TEST=Verify AAC works with patch series and flags enabled

Change-Id: If4368ef04c145604f747baab60e0ab3139e5ec38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4624207
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>