audio_diagonstics: Analyze uptime for CRAS crash
Analyze CRAS and system uptime to determine if CRAS has crashed or not
and print the analysis result for human to read and for FRA to grep.

If system uptime - CRAS uptime > 1 min, then CRAS must have crashed and
restarted.

If CRAS uptime is <= 15 min, count as CRAS recently crashed and print
the result with name: "uptime-cras-recently-crashed"

Otherwise, count as CRAS crashed some time ago and print the result with
name: "uptime-cras-crashed"

Also check if CRAS is not running and print "uptime-cras-not-running"
event.

BUG=b:270292270
TEST=cargo test
TEST=1. Run `restart cras`
     2. `audio_diagnostics` and check "uptime-cras-recently-crashed"
     3. Wait 15 minutes, then run `audio_diagnostics` and check
        "uptime-cras-crashed"
TEST=`stop cras`, then `audio_diagnostics` and check
     "uptime-cras-not-running" event

Change-Id: I63c50abfaf9cf441f74d9dea6dfb490451ea0d27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4622979
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Auto-Submit: Pattara Teerapong <pteerapong@chromium.org>