CRAS: Add switch and TLV bytes config control API via ALSA
Provides read/write API for both switch and TLV bytes typed config
controls. Controls are identified by the name without regard to the card
index. This is under the premise that the target control is on internal
sound card and its name is unique. When there are more than one internal
cards, the interface will iterate on each card index until the target
control is detected by name.

The connected controls will be pushed in a list after use, which can
be restored for the next time read/write the same control. They will be
released on sound card destroyed.

For TLV bytes config write, the premilinary config read is emplaced in
the current design, for the purpose of getting the up-to-date blob
header info.

BUG=b:188647460
TEST=unittest

Change-Id: Ide52bcde2b64c30530257c9f1d0bf7d21e36de9b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2999763
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Pin-chih Lin <johnylin@chromium.org>
Tested-by: Pin-chih Lin <johnylin@chromium.org>
Auto-Submit: Pin-chih Lin <johnylin@chromium.org>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>