[CRAS] Fix regression for not creating USB node for snd-dummy
In the previous change, snd-dummy was not created as an audio node because its parent device is not an actual USB device.
However, we need to use snd-dummy as a USB card simulation in CRAS for testing purposes.
Therefore, we need to make it a special case in order to generate a USB audio node.

BUG=b:292748746
TEST=tast run root@ip audio.USBVolumeGranularity
TEST=USB headset ACM-6005 playback normally

Change-Id: I69806ef1594801c70dbdfe79a4264966a3b909b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4721140
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Ching Yun Chang <whalechang@google.com>