CRAS: prevent assigning NB_MIC when SWB advertised
The BT stack does not provide a convenient way to express all available
HFP codecs. It advertises at most one of WBS or SWB, in which case when
the latter is advertised, CRAS may reject it due to experiment.

Since it is unlikely that there will be headsets supporting SWB but not
WBS, we prevent assigning the node type NB_MIC when SWB is advertised,
(i.e., ignore the flag value).

Note the critical logic determining whether to run SWB lives in
|cras_floss_hfp_start|, which supplies |disabled_codecs| decided by the
flag.

BUG=b:269970706
TEST=Verify HS is not marked as NB_MIC when SWB advertised but disabled

Change-Id: Ic01c965cfd914c9dfcd7b38da2bcaa80a82c2370
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4736897
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>