libcras: Move ScmSocket type into libcras
This type originates in crosvm:
https://github.com/google/crosvm/blob/main/base/src/sys/unix/sock_ctrl_msg.rs

This is a plain copy of that code with a few pieces removed that
are not needed here.
Error types are replaced with nix::Error.

This is a quite complex piece of code that we are duplicating. We
should make use of higher level API's that nix provides here instead
of working directly with libc.

However to unblock migration away from crosvm_base, this CL starts
with a plain copy of the original code.

BUG=b:293488266
TEST=cros build-packages cras-client

Change-Id: Id74edac946164bd9ffff206be93e07648ab8ef7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4750389
Reviewed-by: Judy Hsiao <judyhsiao@google.com>
Tested-by: Dennis Kempin <denniskempin@google.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Dennis Kempin <denniskempin@google.com>