CRAS: prevent reconnection loop in HFP audio disconnection
Some headsets disconnect HFP audio in certain scenarios, such as when
removed from the ears or when in an error state. Usually, when in such
a state, it is possible to fully recover by reconnecting the SCO.

In Floss, when CRAS assumes HFP is started but receives the audio
disconnection callback, we consider it a good timing to attempt a
reconnection to SCO.

However, there is a potential race condition where the reconnection
could trigger the next reconnection, when the disconnection callback
comes after it has been fully reconnected.

In this CL, we add the flag |pending_audio_disconnection| so that
CRAS will not react to the disconnection triggered by itself.

BUG=b:308456681
TEST=Verify HFP reconnection works reliably

Change-Id: Ic66bc5a5d162b39d13874bdbe707c5c172e2dfc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4824548
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>