CRAS: sleep shortly when device is busy on HFP connection
Some peripherals are more likely to hit |EBUSY| when connecting to HFP.
This is relatively easy to reproduce in the scenario when we disconnect
and reconnect to HFP in the shortest possible interval.

When this happens, subsequent requests will fail unless given a
reasonable amount of idle time (i.e., not spamming new requests), which
is not the case as CRAS aggressively switches back and forth considering
the other profile a fallback.

In this CL, we wait for 50ms when the connection to HFP returns EBUSY,
as a mitigation.

BUG=b:295239037
TEST=Verify that Airpods can consistently restart on HFP reconnection

Change-Id: I01b5961f0ea3aaa733b2c420ebb0df34fba0d67a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4825128
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>