CRAS: server_metrics - log HFP SCO reconnection on busy
Sometimes the BT device can be busy when we attempt to connect to SCO.
We have changed the behavior so that in this case we patiently wait for
50ms, in the hope that the reconnection success rate will increase, and
that it won't spam the system as badly if no subsequent reconnection
succeeds.

In this CL, we log the success rate whenever we hit a busy device on HFP
connection.

BUG=b:295239037
TEST=Verify that Airpods can consistently restart on HFP reconnection

Change-Id: I42c8dc8b27ce910cfe26d28457ff057e20779cea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4835817
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>