Use hardware timestamps on ALSA paths
Current calculations of playback and capture timestamps are done with
timestamps based on the current time. This causes issues where the
timestamps obtained by the client are out of order.

As we have hardware timestamps on ALSA paths to ensure the correctness
of the playback and capture timestamps, modify the timestamp calculation
functions to use the hardware timestamps instead.

BUG=b:277656502
TEST=Unit tests

Change-Id: I5da2c316dca8bf7e8d45799db5ec17557aeb2720
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4874527
Commit-Queue: Baili Deng <bailideng@google.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>