[UCM] Enable Software Volume for Chameleon
Derived from b/300626440, tauto audioVolume.usb was broken due to the
uprev'ed Pi.

In order to enable multi-functional USB gadget, we want to uprev
Raspberry Pi from 5.10 to 6.1. However, this uprev will make Pi claim that it has the ability to control volume, so cras will use hardware volume on Pi, which is NOT the expected behavior for audioVolume.usb.

We want to add a registry to UCM, so the sound card simulated by Pi
will be enforced with software volume.

BUG=b:301383800
TEST=local devices with cherry and Pi (Linux 6.1)

Change-Id: I1ef7005a978762eed572d534748c3e8d0a3f0327
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4891635
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Tested-by: ChromeOS Audio Quick Verifier <audio-qv@chromeos-audio-qv.iam.gserviceaccount.com>
Commit-Queue: YJ Lee <yunjunlee@chromium.org>