features: Pull out feature change notification as a callback
We are going to rewrite backend_featured in Rust, which
currently does not have access to C functions in media-sound/adhd.

This makes it impossible for the Rust code to send the feature changed
notification with cras_main_message_send.
Make feature change notification a callback and pass the callback
to the backends so the backend does not need to depend on
cras_main_message.

BUG=b:299569226
TEST=FEATURES='test -usersandbox' emerge-brya adhd

Change-Id: Idd948a07d93aa806a4cbd6b7ff6f6bf6165e4615
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4914699
Reviewed-by: Baili Deng <bailideng@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>