Revert to using current time as timestamp for ALSA output
Based on the diagram of the different time stamps for the playback case
at https://docs.kernel.org/sound/designs/timestamping.html, the analog
time, which is closest to the actual playback time, is calculated using
the app time recorded after writing to the buffer along with the delay
obtained from the ALSA API.

Instead of using the timestamp from the ALSA API, which corresponds to
the DMA time in the diagram, revert to the original implementation of
using the current time as the timestamp. This matches the definition of
the app time in the diagram and results in a more accurate calculation
of the playback time.

BUG=b:303397044
BUG=b:303016089
TEST=tast run $DUT audio.CrasTimestampInOrder.playback on Whiterun

Change-Id: Id68a4125a4980f54c71ccfed510c2f4cf78c0e18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4918855
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>