Add ATLOG event for offset exceeding available frames
Current look through the code show that buffer offsets that are non-zero
are WAI. Therefore, previous patches that limit the number of frames
that are actually written should be enough. However, combined with
underrun writing zeroes without affecting the buffer offset for each
stream, there is still potential of audio issues when the minimum buffer
offset exceeds the number of available frames. This is especially the
case, since previous encounters of the issue would end in crashes, and
we didn't have chances to observe what would happen to the audio in
these cases.

Add ATLOG to help track the events where the minimum buffer offset
exceeds the number of available buffer frames.

BUG=b:302496116
TEST=cq

Change-Id: I6a14ca0eea0098d10943463a5e97fb7c6c24c508
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4931728
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Baili Deng <bailideng@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>