Mark more data as written when stream is draining and empty
Currently, in `get_write_limit`, the write_limit is not limited by
draining streams. This make sense, as we just want to consume all the
data in it, and we know that there won't be further data in the stream.

However, when calculating offset, we are still limited by the draining
stream and its offset to determine how much data is written.

To prevent this, mark streams that are draining and have no data left to
have written more data, up to the write_limit to keep ensure that the
streams with actual data can have their data be properly marked as
written.

BUG=b:306117464
TEST=Have multiple output streams and close them, and check that there
are no `OFFSET_EXCEED_AVAILABLE` events

Change-Id: Ib051a51006e84a84dc7df40d009809f5cda5f563
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4971453
Commit-Queue: Baili Deng <bailideng@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>