CRAS: alsa: Use valid num_channels when using CaptureChannelMap
When CaptureChannelMap is present in the UCM, it is used without
checking the number of channels. Some drivers support a lower number of
channels than what CaptureChannelMap is written for. For example,
CaptureChannelMap can be written with 2 channels in mind, but the device
can also be opened with only 1 channel.

When opening a device with channel layout from CaptureChannelMap, check
that all values are lower than the requested number of channels. If not,
choose a higher number of channels, or fall back to the default channel
layout.

Add ALSAUCMCaptureChannelMapExceedsNumChannels FRA when falling back
to the default channel layout.

BUG=b:302528699
TEST=FEATURES=test emerge-asurada cras_rust
TEST=FEATURES=test emerge-asurada adhd
TEST=`cras_test_client -f S16_LE -c 1 -r 48000 -C /dev/null` on hayato
     with APM enabled and check that CRAS doesn't crash.
TEST=`cras_test_client -f S16_LE -c 1 -r 48000 -C /tmp/test.raw` on
     hayato and check that data is not empty.

Change-Id: I024b3f122db48aca325507f78188b654ea030961
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4988609
Commit-Queue: Pattara Teerapong <pteerapong@chromium.org>
Reviewed-by: Li-Yu Yu <aaronyu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>