CRAS: reset active a2dp device only when inactivating
We used to call |reset_active_device| whenever there is a call to change
the |active| state of an A2DP iodev, as a precautionary measure to
cleanup internal states in the BT stack.

However, this could be an issue when the transition is |active->active|
because by resetting the device there will be a short time where the
socket cannot be written.

Since we have sufficient understanding of the specific internal states
that need to be cleaned up, after this CL we will only reset when
inactivating.

BUG=b:307228349
TEST=Verify bug cannot be reproduced

Change-Id: I76d244a569244d552d3a3131d28f1d533d39b798
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4997244
Auto-Submit: Jeremy Wu <jrwu@google.com>
Reviewed-by: En-Shuo Hsu <enshuo@chromium.org>
Commit-Queue: En-Shuo Hsu <enshuo@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>