CRAS: prevent stacked profile-switch events
When an HFP input device is closed and opened back to back, it would
create two profile-switch events, and that is prone to error.

In this CL, we early return in |bt_io.open_dev| so when there is a
pending profile switch event it would not stack another one when
there is a present queued request.

BUG=b:309547307
TEST=profile-switch and verify Pixel Buds disconnects much less frequently

Change-Id: I01fc3ffddbaad4c1a2588c496f9697763de316d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5004475
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>