CRAS: add new metrics type for SWB
Currently, there are two device types for Bluetooth inputs:
BLUETOOTH and NB_MIC, which map to the metric types WB_MIC
and NB_MIC, respectively.

The device type NB_MIC is special in a way that we want Chrome
to prefer internal mic and thus we distinguish that from
BLUETOOTH (WB_MIC).

Now we have SWB_MIC, we are interested in adding a
corresponding metric type, but it can be treated as the same
device of WB_MIC (BLUETOOTH).

BUG=b:310831425
TEST=FEATURES=test USE=asan emerge-$BOARD adhd

Change-Id: If16bc16e4f0cff0bab033b585bae9dd3755559e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5028533
Tested-by: Jeremy Wu <jrwu@google.com>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>