CRAS: alsa_common - Remove unused flag fully_specified
We have two independent code paths: (a) fully specified UCM and
(b) the legacy, for constructing an ALSA iodev. In (a) we set
this "fully_specified" flag to true, and in (b) we examine whether
this flag is true.

The two paths are independnt, so it makes no sense to have this
flag and related checks at all.  Note that this flag was
unnecessary since it was added and used in commit f57c6c0d24 and
commit d971ddf7d. So we're safe to clean them up.

Also mark the FRA signal USBAudioUCMNoJack and USBAudioUCMWrongJack
as no use.

BUG=None
TEST=None

Change-Id: I0841fc22b7cefd7f0b11e714f76e41d6a91c7362
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5035440
Commit-Queue: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>