ucm: add CRASPlaybackNumberOfVolumeSteps for Zone Vibe Wireless
The Zone Vibe Wireless device correctly reports its volume dB range. However, it incorrectly reports the number of volume steps.

This causes the OS to use an incorrect volume step to control the device, resulting in unexpected behavior. For example, a double volume change event may be sent if the volume does not change significantly enough to trigger an actual volume change on the device (due to the device's granularity of volume steps being larger than the system's).

To resolve this issue, we can overwrite the device volume steps with the CRASPlaybackNumberOfVolumeSteps value.

BUG=b:279803567
TEST=Test with Zone Vibe Wireless volume

Change-Id: I0321ffc2dfdf2c6976ae1d96749412be9fd2a0ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/4914311
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Ching Yun Chang <whalechang@google.com>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>