Verify pipeline channel count
Verify that the channel count for the applied pipeline's source and sink
are the same as the original format for the frames.

This is added to fix crashes, as if the channel counts are different,
this may end up calculating the wrong amount of samples based on the
frames, and result in accessing data out side of the given buffer.

BUG=b:313387350
TEST=Unit tests

Change-Id: I258fde2f2abc67a5c1e2bbb4dc2a54941836be74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5068218
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>