CRAS: assume WBS when SWB supported but disallowed
Floss does not propagate WBS capability of the peer when SWB is
supported, since by the spec when SWB fails to be negotiated, CVSD
follows immediately.

However, as SWB has not launched, it is disallowed by default.
Currently, despite this does not affect the final codec to be used,
this may cause CRAS to assume CVSD will be used before actually starting
the SCO, and result in fallacious device and metric type.

Once SWB has launched, this issue will be gone and thus we apply a
temporary workaround - to assume that if SWB is supported, WBS should
also be supported.

BUG=b:308859926
TEST=connect to a SWB headset and verify there is no NBS warning

Change-Id: Iabfe88c909772c071752a0ed8e9cd0e712e245e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5068217
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Tested-by: Jeremy Wu <jrwu@google.com>
Commit-Queue: Jeremy Wu <jrwu@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>