CRAS: allow fallback to default SBC params in A2DP
While SBC is supposed to be mandatory to be supported and listed in the
available codecs advertised by the peer, we found some cases that this
could not hold, likely due to how AAC is usually the preferred/default
codec to the peers.

If the peer advertises some non-mandatory codec but not SBC, the current
logic will end up in losing A2DP capability. Assuming that SBC should
always be supported but maybe the codec advertisement could not be as
reliable, we add a fallback to a set of default (and must-be-supported)
parameters in such cases.

BUG=b:311233879
TEST=verify A2DP works when SBC is not in codec list

Change-Id: I3f9f8aff4956b53b7c46f25728523af4679c73a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/5076843
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Ching Yun Chang <whalechang@google.com>
Commit-Queue: Ching Yun Chang <whalechang@google.com>